# apps/backend/synapse/contracts/synapse-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synapse Backend API",
    "version": "1.0.0",
    "description": "Shared REST contract implemented by Synapse EE and Cerebro."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/v1/notifications/preferences": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Get my notification preferences (EE Team+)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Preferences", "content": { "application/json": { "schema": { "type": "object" } } } },
          "501": { "description": "Feature not licensed" }
        }
      },
      "put": {
        "tags": ["Notifications"],
        "summary": "Update my notification preferences (EE Team+)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": {
          "200": { "description": "Preferences", "content": { "application/json": { "schema": { "type": "object" } } } },
          "501": { "description": "Feature not licensed" }
        }
      }
    },
    "/v1/notifications/devices": {
      "get": {
        "tags": ["Notifications"],
        "summary": "List my push notification endpoints (EE Team+)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Devices", "content": { "application/json": { "schema": { "type": "object" } } } },
          "501": { "description": "Feature not licensed" }
        }
      },
      "post": {
        "tags": ["Notifications"],
        "summary": "Register a push notification endpoint (EE Team+)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": {
          "201": { "description": "Device registered", "content": { "application/json": { "schema": { "type": "object" } } } },
          "501": { "description": "Feature not licensed" }
        }
      }
    },
    "/v1/notifications/devices/{token_id}": {
      "delete": {
        "tags": ["Notifications"],
        "summary": "Unregister a push notification endpoint (EE Team+)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "token_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "501": { "description": "Feature not licensed" }
        }
      }
    },
    "/v1/notifications/feed": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Recent notification feed for the current user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 50 } }
        ],
        "responses": {
          "200": { "description": "Feed", "content": { "application/json": { "schema": { "type": "object" } } } }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "apiKeyAuth": { "type": "apiKey", "in": "header", "name": "x-api-key" }
    },
    "responses": {
...
      "NotFound": { "description": "Not found" },
...
    },
...
  }
}